+++ src/components/__stories__/buttons.tsx
@@ -2,7 +2,9 @@ import * as React from "react"
 import * as Relay from "react-relay"
 import { storiesOf, action } from "@kadira/storybook"
 
-import Button, { ButtonStyle, ButtonState } from '../button'
+import Button, { ButtonStyle, ButtonState } from '../buttons/default'
+import GhostButton from '../buttons/ghost'
+import InvertedButton from '../buttons/inverted'
 
 storiesOf("Buttons", Button)
     .add("Default Button", () => {
@@ -27,8 +29,8 @@ storiesOf("Buttons", Button)
         return (
             <div>
                 <div>
-                    <Button buttonStyle={ButtonStyle.Inverted}>Inverted Button</Button>
-                    <Button buttonStyle={ButtonStyle.Inverted} disabled>Button</Button>
+                    <InvertedButton>Inverted Button</InvertedButton>
+                    <InvertedButton disabled>Button</InvertedButton>
                 </div>
             </div>
         )
@@ -37,8 +39,8 @@ storiesOf("Buttons", Button)
         return (
             <div>
                 <div>
-                    <Button buttonStyle={ButtonStyle.Ghost}>Ghost Button</Button>
-                    <Button buttonStyle={ButtonStyle.Ghost} disabled>Button</Button>
+                    <GhostButton>Ghost Button</GhostButton>
+                    <GhostButton disabled>Button</GhostButton>
                 </div>
             </div>
         )
